Implement Login screen and update navigation graph
The navigation graph has been updated to include a login screen, which is now the start destination.
The login screen is implemented in `LoginScreen` and can be accessed through the `Routes.Login` route.
The `NavHost` now starts at the `Routes.Login` route.
`MateriasScreen` and `TareasScreen` components are modified to include `authViewModel`